src/mcp-server/src/server.py
"""
MCP Server - Centralized Tool Registry
Exposes enterprise-grade tools that any agent in the system can utilize
"""

import asyncio
import json
import logging
import uuid
from datetime import datetime
from typing import Any, Dict, List, Optional

from adk import MCPToolServer, FunctionTool
from adk_shared.observability import get_tracer, trace_tool_call
from adk_shared.security import authenticate_request
import yaml


class DatabaseQueryTool(FunctionTool):
    """Tool for querying enterprise databases"""
    
    name = "query_database"
    description = "Execute SQL queries against enterprise databases"
    
    async def execute(self, query: str, database: str = "default", **kwargs) -> Dict[str, Any]:
        """Execute database query with observability and security"""
        transaction_id = str(uuid.uuid4())
        
        with trace_tool_call("database_query", transaction_id, {"query": query, "database": database}):
            try:
                # Authenticate the request
                if not authenticate_request(kwargs.get('auth_token')):
                    raise PermissionError("Unauthorized database access")
                
                # Mock database query - replace with actual implementation
                result = {
                    "transaction_id": transaction_id,
                    "query": query,
                    "database": database,
                    "rows": [
                        {"id": 1, "name": "Sample Record 1", "value": 100},
                        {"id": 2, "name": "Sample Record 2", "value": 200}
                    ],
                    "row_count": 2,
                    "execution_time_ms": 150,
                    "timestamp": datetime.now().isoformat()
                }
                
                logging.info(f"Database query executed: {transaction_id}")
                return result
                
            except Exception as e:
                logging.error(f"Database query failed: {transaction_id} - {str(e)}")
                raise


class DocumentSearchTool(FunctionTool):
    """Tool for searching enterprise documents"""
    
    name = "search_documents"
    description = "Search through enterprise document repositories"
    
    async def execute(self, query: str, repository: str = "default", limit: int = 10, **kwargs) -> Dict[str, Any]:
        """Search documents with observability and security"""
        transaction_id = str(uuid.uuid4())
        
        with trace_tool_call("document_search", transaction_id, {"query": query, "repository": repository}):
            try:
                # Authenticate the request
                if not authenticate_request(kwargs.get('auth_token')):
                    raise PermissionError("Unauthorized document access")
                
                # Mock document search - replace with actual implementation
                result = {
                    "transaction_id": transaction_id,
                    "query": query,
                    "repository": repository,
                    "documents": [
                        {
                            "id": "doc_001",
                            "title": "Enterprise Architecture Guidelines",
                            "snippet": "This document outlines best practices...",
                            "score": 0.95,
                            "last_modified": "2024-01-15T10:30:00Z"
                        },
                        {
                            "id": "doc_002",
                            "title": "API Security Standards",
                            "snippet": "Security requirements for all APIs...",
                            "score": 0.87,
                            "last_modified": "2024-01-10T14:20:00Z"
                        }
                    ],
                    "total_results": 2,
                    "search_time_ms": 75,
                    "timestamp": datetime.now().isoformat()
                }
                
                logging.info(f"Document search executed: {transaction_id}")
                return result
                
            except Exception as e:
                logging.error(f"Document search failed: {transaction_id} - {str(e)}")
                raise


class AnalyticsTool(FunctionTool):
    """Tool for running analytics and generating insights"""
    
    name = "run_analytics"
    description = "Execute analytics queries and generate business insights"
    
    async def execute(self, analysis_type: str, parameters: Dict[str, Any], **kwargs) -> Dict[str, Any]:
        """Run analytics with observability and security"""
        transaction_id = str(uuid.uuid4())
        
        with trace_tool_call("analytics", transaction_id, {"type": analysis_type, "params": parameters}):
            try:
                # Authenticate the request
                if not authenticate_request(kwargs.get('auth_token')):
                    raise PermissionError("Unauthorized analytics access")
                
                # Mock analytics - replace with actual implementation
                result = {
                    "transaction_id": transaction_id,
                    "analysis_type": analysis_type,
                    "parameters": parameters,
                    "insights": [
                        {
                            "metric": "revenue_growth",
                            "value": 15.3,
                            "trend": "increasing",
                            "confidence": 0.92
                        },
                        {
                            "metric": "user_engagement",
                            "value": 78.5,
                            "trend": "stable",
                            "confidence": 0.88
                        }
                    ],
                    "processing_time_ms": 2100,
                    "timestamp": datetime.now().isoformat()
                }
                
                logging.info(f"Analytics executed: {transaction_id}")
                return result
                
            except Exception as e:
                logging.error(f"Analytics failed: {transaction_id} - {str(e)}")
                raise


# MCP Server main class
class EnterpriseToolServer(MCPToolServer):
    """Enterprise MCP Tool Server with governance and observability"""
    
    def __init__(self, config_path: str = "config/tools.yaml"):
        super().__init__()
        self.config = self._load_config(config_path)
        self._register_tools()
        
    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """Load tool configuration from YAML"""
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    
    def _register_tools(self):
        """Register all available tools"""
        # Register core tools
        self.register_tool(DatabaseQueryTool())
        self.register_tool(DocumentSearchTool())
        self.register_tool(AnalyticsTool())
        
        logging.info("Enterprise tools registered successfully")
    
    async def start_server(self, host: str = "0.0.0.0", port: int = 8000):
        """Start the MCP server with observability"""
        tracer = get_tracer("mcp-server")
        
        with tracer.start_as_current_span("mcp_server_startup"):
            logging.info(f"Starting Enterprise MCP Server on {host}:{port}")
            await super().start_server(host, port)
